updated all todo resolvers and services with current logged in user

+++ server/src/todos/todo.service.ts
@@ -3,6 +3,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Todo } from './todo.entity';
 import { UserService } from '../user/user.service';
+import { User } from '../user/user.entity';
 import { InputTodo, DateCompletedTodo } from './input/todo.input';
 
 
@@ -15,39 +16,44 @@ export class TodoService {
     ) {}
 
     // Find all Todos that belong to a user
-    findAllTodos(userId: string): Promise<Todo[]> {
-        return this.todoRepository.find({ where: { user: userId } });
+    findAllTodos(user: User): Promise<Todo[]> {
+        return this.todoRepository.find({ where: { user } });
     }
 
     // Get all Todos for a user that are currently open
-    allOpenTodos(userId: string): Promise<Todo[]> {
-        return this.todoRepository.find({ where: { user: userId, status: 'open' }, order: { dueDate: "DESC" } });
+    allOpenTodos(user: User): Promise<Todo[]> {
+        return this.todoRepository.find({ where: { user, status: 'open' }, order: { dueDate: "DESC" } });
     }
 
     // Get a specific user for a todo
     async findOneTodo(id: string, userId: string): Promise<Todo> {
         const todo = await this.todoRepository.findOne({ where: { id, user: userId }, order: { dueDate: "DESC" }} );
         if (userId !== todo.user.id) {
-            throw new Error("You don't have access to this Todo")
+            throw new Error("You don't have access to this Todo");
         }
         return todo;
     } 
 
     // Delete a specific Todo
-    async removeTodo(id: string): Promise<void> {
+    async removeTodo(id: string, userId: string): Promise<void> {
+        const todo = await this.todoRepository.findOne({ where: { id } });
+        if (userId !== todo.user.id) {
+            throw new Error("You don't have access to this Todo");
+        }
         await this.todoRepository.delete(id);
     }
 
     // Need to make sure FK to logged in user is added in
-    async createTodo(data: InputTodo): Promise<Todo> {
-        const foundTodo = await this.todoRepository.findOne({ title: data.title })
-        if (foundTodo) throw new Error("This title already exists, a unique title is required")
+    async createTodo(data: InputTodo, user: User): Promise<Todo> {
+        const foundTodo = await this.todoRepository.findOne({ where: { user, title: data.title } });
+        if (foundTodo) throw new Error("This title already exists, a unique title is required");
 
         const todo = new Todo();
         todo.title = data.title;
         todo.body = data.body;
         todo.dueDate = new Date(data.dueDate).toISOString();
         todo.status = 'open';
+        todo.user = user;
 
         await this.todoRepository.save(todo);
 
@@ -55,8 +61,11 @@ export class TodoService {
     }
 
     // Add completed date to Todo
-    async addDateCompleted(id: string, data: DateCompletedTodo): Promise<Todo> {
+    async addDateCompleted(id: string, data: DateCompletedTodo, userId: string): Promise<Todo> {
         const todo = await this.todoRepository.findOne(id);
+        if(userId !== todo.user.id) {
+            throw new Error("You do not have access to this Todo");
+        }
         todo.dateCompleted = new Date(data.dateCompleted).toISOString();
         todo.status = 'complete';
 
@@ -66,8 +75,11 @@ export class TodoService {
     }
 
     // Update Todo
-    async updateTodo(id: string, data: InputTodo): Promise<Todo> {
+    async updateTodo(id: string, data: InputTodo, userId: string): Promise<Todo> {
         const todo = await this.todoRepository.findOne(id);
+        if (userId !== todo.user.id) {
+            throw new Error("You do not have access to this Todo");
+        }
         return this.todoRepository.save({
             ...todo,
             ...data
